+++ src/Models/SystemData.h
@@ -68,6 +68,10 @@ struct SystemFirmware
     char buildTime[Model::BUILD_TIME_LENGTH] {};
 
     char resetReason[Model::RESET_REASON_LENGTH] {};
+
+    char resetInfo[192] {};
+
+    bool exceptionReset = false;
 };
 
 //=============================================================================